[IMP] hetzner: Refactor
